pyproject.toml
[project]
name = "apex-ledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "anthropic>=0.86.0",
    "asyncpg==0.29.0",
    "docling>=2.82.0",
    "fastapi>=0.135.1",
    "langgraph==1.0.10",
    "mineru>=2.7.6",
    "openai==2.24.0",
    "pandas>=3.0.1",
    "pdfplumber==0.11.9",
    "pydantic==2.7.4",
    "pydantic-settings>=2.13.1",
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.3",
    "scikit-learn>=1.8.0",
    "umap-learn>=0.5.11",
    "uvicorn[standard]>=0.42.0",
]
